requirements.txt
streamlit
feedparser
google-genai
pandas
plotly
beautifulsoup4
python-dotenv
watchdog
python-dateutil
